Create fizzbuzz.c
My first working attempt at FizzBuzz without looking at any answers first. 

I created a boolean flag ("normal") to determine if I need to print a number instead of Fizz and/or Buzz. This works well, but it feels like I should be able to get around this with better if statement structure.